Toolchain: removed redundant downloads

+++ BuildScripts/Toolchain.mk
@@ -5,14 +5,14 @@ TOOLCHAIN_PREFIX:=$(TOOLCHAIN_PATH)/$(TARGET)
 export PATH:=$(PATH):$(TOOLCHAIN_PREFIX)/bin
 
 all: toolchain
-toolchain: toolchain_binutils toolchain_binutils toolchain_gcc
+toolchain: toolchain_binutils toolchain_gcc
 toolchain_binutils: | $(TOOLCHAIN_PREFIX)/bin/$(TARGET)-ld
 toolchain_gcc: | $(TOOLCHAIN_PREFIX)/bin/$(TARGET)-gcc
 
 .PHONY: toolchain clean-toolchain
 
 $(TOOLCHAIN_PREFIX)/bin/$(TARGET)-ld: | $(TOOLCHAIN_PATH)
-	wget $(BINUTILS_URL) -O $(TOOLCHAIN_PATH)/binutils.tar.xz
+	if [ ! -f "$(TOOLCHAIN_PATH)/binutils.tar.xz" ]; then wget $(BINUTILS_URL) -O $(TOOLCHAIN_PATH)/binutils.tar.xz; fi
 	mkdir -p $(TOOLCHAIN_PATH)/binutils-src
 	tar -xf $(TOOLCHAIN_PATH)/binutils.tar.xz -C $(TOOLCHAIN_PATH)/binutils-src --strip-components 1
 	rm -rf $(TOOLCHAIN_PATH)/binutils-build && mkdir -p $(TOOLCHAIN_PATH)/binutils-build && cd $(TOOLCHAIN_PATH)/binutils-build && \
@@ -21,7 +21,7 @@ $(TOOLCHAIN_PREFIX)/bin/$(TARGET)-ld: | $(TOOLCHAIN_PATH)
 	$(MAKE) -C $(TOOLCHAIN_PATH)/binutils-build install
 
 $(TOOLCHAIN_PREFIX)/bin/$(TARGET)-gcc: | $(TOOLCHAIN_PATH)
-	wget $(GCC_URL) -O $(TOOLCHAIN_PATH)/gcc.tar.gz
+	if [ ! -f "$(TOOLCHAIN_PATH)/gcc.tar.gz" ]; then wget $(GCC_URL) -O $(TOOLCHAIN_PATH)/gcc.tar.gz; fi
 	mkdir -p $(TOOLCHAIN_PATH)/gcc-src
 	tar -xf $(TOOLCHAIN_PATH)/gcc.tar.gz -C $(TOOLCHAIN_PATH)/gcc-src --strip-components 1
 	rm -rf $(TOOLCHAIN_PATH)/gcc-build && mkdir -p $(TOOLCHAIN_PATH)/gcc-build && cd $(TOOLCHAIN_PATH)/gcc-build && \
